requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via httpx
arrow==1.4.0
    # via
    #   -r requirements.in
    #   cookiecutter
asgiref==3.11.0
    # via
    #   django
    #   django-browser-reload
    #   django-cors-headers
authlib==1.6.6
    # via
    #   -r requirements.in
    #   safety
bandit==1.9.2
    # via -r requirements.in
binaryornot==0.4.4
    # via cookiecutter
build==1.3.0
    # via pip-tools
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   cookiecutter
    #   nltk
    #   pip-tools
    #   safety
    #   typer
cookiecutter==2.6.0
    # via -r requirements.in
cryptography==46.0.3
    # via authlib
django==6.0
    # via
    #   -r requirements.in
    #   django-browser-reload
    #   django-cors-headers
    #   django-csp
    #   django-ranged-response
    #   django-simple-captcha
    #   django-taggit
    #   django-tailwind
    #   djangorestframework
django-browser-reload==1.21.0
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-csp==4.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-ranged-response==0.2.0
    # via
    #   -r requirements.in
    #   django-simple-captcha
django-ratelimit==4.1.0
    # via -r requirements.in
django-simple-captcha==0.6.3
    # via -r requirements.in
django-taggit==6.1.0
    # via -r requirements.in
django-tailwind==4.4.2
    # via -r requirements.in
django-webpack-loader==3.2.3
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
dparse==0.6.4
    # via
    #   safety
    #   safety-schemas
filelock==3.20.2
    # via safety
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via safety
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   cookiecutter
    #   safety
joblib==1.5.3
    # via nltk
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
marshmallow==4.2.0
    # via safety
mdurl==0.1.2
    # via markdown-it-py
nltk==3.9.2
    # via safety
packaging==25.0
    # via
    #   build
    #   django-csp
    #   dparse
    #   gunicorn
    #   safety
    #   safety-schemas
pillow==12.1.0
    # via
    #   -r requirements.in
    #   django-simple-captcha
pip==25.3
    # via pip-tools
pip-tools==7.5.2
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   safety
    #   safety-schemas
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pymemcache==4.0.0
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytailwindcss==0.3.0
    # via django-tailwind
python-dateutil==2.9.0.post0
    # via arrow
python-slugify==8.0.4
    # via
    #   -r requirements.in
    #   cookiecutter
pyyaml==6.0.3
    # via
    #   bandit
    #   cookiecutter
regex==2025.11.3
    # via nltk
requests==2.32.5
    # via
    #   -r requirements.in
    #   cookiecutter
    #   safety
rich==14.2.0
    # via
    #   bandit
    #   cookiecutter
    #   typer
ruamel-yaml==0.19.1
    # via
    #   safety
    #   safety-schemas
safety==3.7.0
    # via -r requirements.in
safety-schemas==0.0.16
    # via safety
setuptools==80.9.0
    # via pip-tools
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
stevedore==5.6.0
    # via bandit
tenacity==9.1.2
    # via safety
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.3
    # via safety
tqdm==4.67.1
    # via nltk
typer==0.21.0
    # via safety
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   safety
    #   safety-schemas
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via arrow
urllib3==2.6.2
    # via requests
wheel==0.45.1
    # via pip-tools
whitenoise==6.11.0
    # via -r requirements.in
